% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whensxmas.R
\name{whensxmas}
\alias{whensxmas}
\title{Days Until Christmas}
\usage{
whensxmas(...)
}
\arguments{
\item{...}{Additional arguments. Currently not implemented.}
}
\value{
Returns a message with the number of days left until Christmas.
}
\description{
A simple function to tell you how many days are left until the next Christmas.
}
\details{
The description says it all - plain and simple!
}
\examples{
{
whensxmas()
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
